% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{coef.tune_elastic_net}
\alias{coef.tune_elastic_net}
\title{Extract coefficients from tune_elastic_net object}
\usage{
\method{coef}{tune_elastic_net}(object, s = "lambda.min", ...)
}
\arguments{
\item{object}{A \code{tune_elastic_net} object.}

\item{s}{Value(s) of lambda at which to extract coefficients. Default is "lambda.min".
Can also be "lambda.1se" or numeric value(s).}

\item{...}{Additional arguments passed to \code{coef.cv.glmnet}.}
}
\value{
A sparse matrix of coefficients.
}
\description{
Extract coefficients from tune_elastic_net object
}
