% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{lrt}
\alias{lrt}
\title{Likelihood ratio test for nested models}
\usage{
lrt(model_small, model_large)
}
\arguments{
\item{model_small}{The simpler (restricted) model.}

\item{model_large}{The more complex (full) model.}
}
\value{
A tibble with columns: statistic, df, p_value.
}
\description{
Performs a likelihood ratio test comparing two nested models.
}
\examples{
m1 <- lm(mpg ~ wt, mtcars)
m2 <- lm(mpg ~ wt + hp, mtcars)
lrt(m1, m2)
}
