% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{poly_reg}
\alias{poly_reg}
\title{Fit polynomial regression}
\usage{
poly_reg(data, response, predictor, degree)
}
\arguments{
\item{data}{A data frame.}

\item{response}{Name of the response variable (character).}

\item{predictor}{Name of the predictor variable (character).}

\item{degree}{Polynomial degree (integer).}
}
\value{
A fitted \code{lm} object.
}
\description{
Fits a polynomial regression model of specified degree.
}
\examples{
poly_reg(mtcars, "mpg", "wt", degree = 3)
}
