% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{predict.tune_elastic_net}
\alias{predict.tune_elastic_net}
\title{Predict method for tune_elastic_net objects}
\usage{
\method{predict}{tune_elastic_net}(object, newx, s = "lambda.min", ...)
}
\arguments{
\item{object}{A \code{tune_elastic_net} object.}

\item{newx}{Matrix of new predictor values for prediction.}

\item{s}{Value(s) of lambda at which to make predictions. Default is "lambda.min".
Can also be "lambda.1se" or numeric value(s).}

\item{...}{Additional arguments passed to \code{predict.cv.glmnet}.}
}
\value{
Vector or matrix of predictions.
}
\description{
Predict method for tune_elastic_net objects
}
