% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{print.tune_elastic_net}
\alias{print.tune_elastic_net}
\title{Print method for tune_elastic_net objects}
\usage{
\method{print}{tune_elastic_net}(x, ...)
}
\arguments{
\item{x}{A \code{tune_elastic_net} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for tune_elastic_net objects
}
