% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{train_test_split}
\alias{train_test_split}
\title{Split data into training and test sets}
\usage{
train_test_split(data, prop = 0.5)
}
\arguments{
\item{data}{A data frame to split.}

\item{prop}{Proportion of data for training (default 0.5).}
}
\value{
A list with \code{train} and \code{test} data frames.
}
\description{
Randomly splits a data frame into training and test subsets.
}
\examples{
split <- train_test_split(mtcars, prop = 0.7)
split$train
split$test
}
